Begin moving project source code into root dir
[electron-webpack][0] has a very specific project directory structure
which must be used, if you want to use electron-webpack for building the
project. (And you should want to use it, as it does most of the hard
work of configuring webpack for you.)

Begin moving some of the source code from `app/` directory into the
toplevel `src/` directory so that the we can run the dev-mode/compiler.

[0]: https://webpack.electron.build/project-structure