Introduce window.innerWidth and window.innerHeight
Global variable `window` isn't currently usable because type
declarations for it are missing. Because of this, size of the game had
to be hardcoded, because it couldn't be read from `window.innerWidth`
and `window.innerHeight`.

As a temporary solution introduce global variable `window` with
`innerWidth` and `innerHeight` properties. As a long term solution, we
are going to need type declarations for the browser API at some point.

And also turn `StatesBoot` into proper class which is introduced as
instance of that class into `game.states` without using the unnecessary
`prep_state` function at all.

Fixes #134